ATR-304: moved suppress message generation into SuppressMessage own declaration
