+++ src/Acuminator/Acuminator.Utilities/DiagnosticSuppression/SuppressMessage.cs
@@ -1,5 +1,11 @@
-﻿using Acuminator.Utilities.Common;
-using System;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Acuminator.Utilities.Common;
 
 namespace Acuminator.Utilities.DiagnosticSuppression
 {
@@ -8,6 +14,14 @@ namespace Acuminator.Utilities.DiagnosticSuppression
 	/// </summary>
 	public readonly struct SuppressMessage : IEquatable<SuppressMessage>, IComparable<SuppressMessage>
     {
+		private static HashSet<SyntaxKind> _targetKinds = new HashSet<SyntaxKind>(new[] {
+			SyntaxKind.ClassDeclaration,
+			SyntaxKind.MethodDeclaration,
+			SyntaxKind.ConstructorDeclaration,
+			SyntaxKind.PropertyDeclaration,
+			SyntaxKind.FieldDeclaration
+		});
+
 		/// <summary>
 		/// Suppressed diagnostic Id
 		/// </summary>
@@ -109,5 +123,90 @@ namespace Acuminator.Utilities.DiagnosticSuppression
 
             return 0;
         }
-    }
+
+		public static (string Assembly, SuppressMessage Message) GetSuppressionInfo(SemanticModel semanticModel, Diagnostic diagnostic,
+																					 CancellationToken cancellation = default)
+		{
+			cancellation.ThrowIfCancellationRequested();
+
+			if (semanticModel == null || diagnostic?.Location == null)
+				return (null, default);
+
+			var rootNode = semanticModel.SyntaxTree.GetRoot(cancellation);
+			if (rootNode == null)
+				return (null, default);
+
+			var diagnosticNode = rootNode.FindNode(diagnostic.Location.SourceSpan);
+			if (diagnosticNode == null)
+				return (null, default);
+
+			var targetNode = FindTargetNode(diagnosticNode);
+			if (targetNode == null)
+				return (null, default);
+
+			var targetSymbol = semanticModel.GetDeclaredSymbol(targetNode, cancellation);
+			if (targetSymbol == null)
+				return (null, default);
+
+			var assemblyName = targetSymbol.ContainingAssembly?.Name;
+			if (string.IsNullOrEmpty(assemblyName))
+				return (null, default);
+
+			var target = targetSymbol.ToDisplayString();
+			string syntaxNodeString = GetSyntaxNodeStringForSuppressionMessage(diagnosticNode,
+																			   diagnosticPosition: diagnostic.Location.SourceSpan.Start);
+			var message = new SuppressMessage(diagnostic.Id, target, syntaxNodeString);
+			return (assemblyName, message);
+		}
+
+		private static SyntaxNode FindTargetNode(SyntaxNode node)
+		{
+			if (node == null)
+				return null;
+
+			var targetNode = node
+				.AncestorsAndSelf()
+				.Where(a => _targetKinds.Contains(a.Kind()))
+				.FirstOrDefault();
+
+			if (!(targetNode is FieldDeclarationSyntax fieldDeclaration))
+				return targetNode;
+			else if (fieldDeclaration.Declaration == null)
+				return null;
+
+			SeparatedSyntaxList<VariableDeclaratorSyntax> declaredVariables = fieldDeclaration.Declaration.Variables;
+
+			switch (declaredVariables.Count)
+			{
+				case 0:
+					return null;
+				case 1:
+					return declaredVariables[0];
+				default:
+					return declaredVariables.FirstOrDefault(variableDeclarator => variableDeclarator.Contains(node));
+			}
+		}
+
+		private static string GetSyntaxNodeStringForSuppressionMessage(SyntaxNode diagnosticNode, int diagnosticPosition)
+		{
+			// Try to obtain token in case of member declaration syntax as we do not want to store the text of the entire declaration node
+			SyntaxToken? token = null;
+
+			if (diagnosticNode is MemberDeclarationSyntax memberDeclaration)
+			{
+				try
+				{
+					token = memberDeclaration.FindToken(diagnosticPosition);
+				}
+				catch (ArgumentOutOfRangeException)
+				{
+					token = null;
+				}
+			}
+
+			return token != null
+				? token.ToString()
+				: diagnosticNode.ToString().Replace("\r", "");  // Replace \r symbol as XDocument does not preserve it in suppression file
+		}
+	}
 }
