+++ src/Acuminator/Acuminator.Utilities/DiagnosticSuppression/SuppressionManager.cs
@@ -13,15 +13,6 @@ namespace Acuminator.Utilities.DiagnosticSuppression
 	public class SuppressionManager
 	{
 		private readonly ConcurrentDictionary<string, SuppressionFile> _fileByAssembly = new ConcurrentDictionary<string, SuppressionFile>();
-
-		private static HashSet<SyntaxKind> _targetKinds = new HashSet<SyntaxKind>(new[] {
-			SyntaxKind.ClassDeclaration,
-			SyntaxKind.MethodDeclaration,
-			SyntaxKind.ConstructorDeclaration,
-			SyntaxKind.PropertyDeclaration,
-			SyntaxKind.FieldDeclaration
-		});
-
 		private readonly ISuppressionFileSystemService _fileSystemService;
 
 		private static SuppressionManager Instance { get; set; }
@@ -105,7 +96,7 @@ namespace Acuminator.Utilities.DiagnosticSuppression
 		{
 			cancellation.ThrowIfCancellationRequested();
 
-			var (assembly, message) = GetSuppressionInfo(semanticModel, diagnostic, cancellation);
+			var (assembly, message) = SuppressMessage.GetSuppressionInfo(semanticModel, diagnostic, cancellation);
 
 			if (assembly == null)
 			{
@@ -183,91 +174,6 @@ namespace Acuminator.Utilities.DiagnosticSuppression
 			}
 
 			reportDiagnostic(diagnostic);
-		}
-
-		public static (string Assembly, SuppressMessage Message) GetSuppressionInfo(SemanticModel semanticModel, Diagnostic diagnostic,
-																					 CancellationToken cancellation = default)
-		{
-			cancellation.ThrowIfCancellationRequested();
-
-			if (semanticModel == null || diagnostic?.Location == null)
-				return (null, default);
-
-			var rootNode = semanticModel.SyntaxTree.GetRoot(cancellation);
-			if (rootNode == null)
-				return (null, default);
-
-			var diagnosticNode = rootNode.FindNode(diagnostic.Location.SourceSpan);
-			if (diagnosticNode == null)
-				return (null, default);
-
-			var targetNode = FindTargetNode(diagnosticNode);
-			if (targetNode == null)
-				return (null, default);
-
-			var targetSymbol = semanticModel.GetDeclaredSymbol(targetNode, cancellation);
-			if (targetSymbol == null)
-				return (null, default);
-
-			var assemblyName = targetSymbol.ContainingAssembly?.Name;
-			if (string.IsNullOrEmpty(assemblyName))
-				return (null, default);
-
-			var target = targetSymbol.ToDisplayString();
-			string syntaxNodeString = GetSyntaxNodeStringForSuppressionMessage(diagnosticNode, 
-																			   diagnosticPosition: diagnostic.Location.SourceSpan.Start);
-			var message = new SuppressMessage(diagnostic.Id, target, syntaxNodeString);
-			return (assemblyName, message);
-		}
-
-		private static SyntaxNode FindTargetNode(SyntaxNode node)
-		{
-			if (node == null)
-				return null;
-
-			var targetNode = node
-				.AncestorsAndSelf()
-				.Where(a => _targetKinds.Contains(a.Kind()))
-				.FirstOrDefault();
-
-			if (!(targetNode is FieldDeclarationSyntax fieldDeclaration))
-				return targetNode;
-			else if (fieldDeclaration.Declaration == null)
-				return null;
-
-			SeparatedSyntaxList<VariableDeclaratorSyntax> declaredVariables = fieldDeclaration.Declaration.Variables;
-
-			switch (declaredVariables.Count)
-			{
-				case 0:
-					return null;
-				case 1:
-					return declaredVariables[0];
-				default:
-					return declaredVariables.FirstOrDefault(variableDeclarator => variableDeclarator.Contains(node));
-			}
-		}
-
-		private static string GetSyntaxNodeStringForSuppressionMessage(SyntaxNode diagnosticNode, int diagnosticPosition)
-		{
-			// Try to obtain token in case of member declaration syntax as we do not want to store the text of the entire declaration node
-			SyntaxToken? token = null;
-
-			if (diagnosticNode is MemberDeclarationSyntax memberDeclaration)
-			{
-				try
-				{
-					token = memberDeclaration.FindToken(diagnosticPosition);
-				}
-				catch (ArgumentOutOfRangeException)
-				{
-					token = null;
-				}
-			}
-
-			return token != null 
-				? token.ToString()
-				: diagnosticNode.ToString().Replace("\r", "");	// Replace \r symbol as XDocument does not preserve it in suppression file
 		}
 	}
 }
